Refactor losses and remove unused elbo function
